requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
ultralytics==8.0.202
numpy==1.24.3
opencv-python==4.8.1.78
Pillow==10.0.1
python-multipart==0.0.6